# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Varterm Text-to-Speech API",
    "description": "Convert text to natural-sounding speech using premium AI voices. Varterm provides high-quality text-to-speech capabilities with multiple voice options.",
    "version": "1.0.0",
    "contact": {
      "name": "Varterm",
      "url": "https://varterm.com"
    }
  },
  "servers": [
    {
      "url": "https://varterm.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/tts": {
      "get": {
        "operationId": "listVoices",
        "summary": "List available voices",
        "description": "Get a list of all available text-to-speech voices with their IDs and descriptions",
        "responses": {
          "200": {
            "description": "List of available voices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "voices": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Voice"
                      }
                    },
                    "default_voice": {
                      "type": "string",
                      "description": "Default voice ID"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "textToSpeech",
        "summary": "Convert text to speech",
        "description": "Generate natural-sounding audio from text using AI voices. Returns an MP3 audio file.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio file generated successfully",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Voice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique voice identifier"
          },
          "name": {
            "type": "string",
            "description": "Voice name (e.g., Sarah, Charlie)"
          },
          "description": {
            "type": "string",
            "description": "Voice characteristics"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female"]
          }
        }
      },
      "TTSRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string",
            "description": "The text to convert to speech (max 5000 characters)",
            "maxLength": 5000
          },
          "voice": {
            "type": "string",
            "description": "Voice name (e.g., 'Sarah', 'Charlie'). Use listVoices to see options."
          },
          "voiceId": {
            "type": "string",
            "description": "Voice ID (alternative to voice name)"
          },
          "speed": {
            "type": "number",
            "description": "Speech speed (0.5 to 2.0, default 1.0)",
            "minimum": 0.5,
            "maximum": 2.0,
            "default": 1.0
          },
          "stability": {
            "type": "number",
            "description": "Voice stability (0 to 1, default 0.5)",
            "minimum": 0,
            "maximum": 1,
            "default": 0.5
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    }
  }
}